Update index.html
First go was to keep all css in separate file, but had difficulty with the image background with text overlay.  So I went with inline styling.
The advised typeface (Helvetica) is not the same as the provided model and affected the layout.
Still need to add the text color for pricing.
Looks ok.  There is much to learn.